feat: merge flip and autoPlacement into adaptivePlacement
fix import

fix types

fix exports

fix react dom tests

fix tests
