+++ packages/core/src/index.ts
@@ -1,9 +1,12 @@
 export {computePosition} from './computePosition';
 export {detectOverflow} from './detectOverflow';
 
+export {
+  adaptivePlacement,
+  fallback,
+  bestFit,
+} from './middleware/adaptivePlacement';
 export {arrow} from './middleware/arrow';
-export {autoPlacement} from './middleware/autoPlacement';
-export {flip} from './middleware/flip';
 export {hide} from './middleware/hide';
 export {offset} from './middleware/offset';
 export {shift, limitShift} from './middleware/shift';
