+++ packages/core/test/middleware/autoPlacement.test.ts
@@ -1,4 +1,4 @@
-import {getPlacementList} from '../../src/middleware/autoPlacement';
+import {getPlacementList} from '../../src/middleware/adaptivePlacement';
 
 test('base placement', () => {
   expect(
